pyproject.toml
[tool.poetry]
name = "task manager api"
version = "0.1.0"
description = "API that helps manage and track task lifecycle"
authors = ["rifatrakib <Rakib.1508@outlook.com>"]
license = "Apache License version 2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.95.1"}
SQLAlchemy = {extras = ["asyncion"], version = "^2.0.12"}
pydantic = {extras = ["dotenv", "email"], version = "^1.10.7"}
asyncpg = "^0.27.0"
alembic = "^1.10.4"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
ruff = "^0.0.264"
pytest = "^7.3.1"
coverage = "^7.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
)/
'''

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10

[tool.isort]
atomic = true
include_trailing_comma = true
profile = "black"
line_length = 120
skip_gitignore = true

[tool.ruff]
line-length = 120
exclude = [".git", "__pycache__", "dist"]
target-version = "py39"
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.mccabe]
max-complexity = 10
